Hotel_booking_system/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==2.2.8
django-celery-results==2.4.0
django-cors-headers==3.11.0
django-crontab==0.7.1
djangorestframework==3.12.4
future==0.18.2
kombu==5.2.4
packaging==21.3
passlib==1.7.2
Pillow==9.0.0
prompt-toolkit==3.0.31
pyparsing==3.0.9
pytz==2021.3
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
